Did task service

+++ src/main/java/com/test/taskmanager/dto/task/TasksDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 @NoArgsConstructor
 public class TasksDTO {
 
-    private Long count;
+    private Integer count;
     private List<TaskDTO> tasks;
 
 }
